Fix return datatype issue for buffer

+++ temp_sensors.py
@@ -51,7 +51,7 @@ class Buffer:
         d = np.abs(data - np.median(data))
         mdev = np.median(d)
         s = d / (mdev if mdev else 1.)
-        return data[s < m].mean()
+        return float(data[s < m].mean())
 
     
     def reset(self) -> None:
